+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-0.0.0 (16/11/2019)
+0.0.2 (05/12/2019)
 ------------------
-* Init commit
+* Added models operations for ServerStatus
+* Added unit and integration tests for ServerStatus
 
 0.0.1 (29/11/2019)
 ------------------
 * Added models operations: Admin, Ip
 * Added hound and flake8 configuration
 * added TravisCI configuration
+
+0.0.0 (16/11/2019)
+------------------
+* Init commit
+